Add files via upload
"The Employee Management System is a web application developed using Django, HTML, and CSS. It provides a user-friendly interface to efficiently manage employees within an organization. The system allows users to view a comprehensive list of all employees, including their details such as name, position, and contact information. Users can easily add new employees to the system by filling in the required information. Additionally, the system enables users to remove employees from the database if needed. To enhance usability, the application also includes a filtering feature, allowing users to search for specific employees based on different criteria. With its intuitive interface and functionality, the Employee Management System simplifies the process of organizing and maintaining employee records in a hassle-free manner."